projects/project3/examples/중/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
# Gemini API 실제 연동 시 주석 해제
# google-generativeai>=0.5.0
